made it easy to choose RGB or black and white

+++ code/metrics.py
@@ -21,8 +21,8 @@ def specificity(y_true, y_pred):
     # n = tf.keras.metrics.TrueNegatives(threshold)
     # n.update_state(y_true,y_pred)
     # tn = m.result().numpy()
-    y_true = tf.argmax(y_true, axis=-1)
-    y_pred = tf.argmax(y_pred, axis=-1)
+    y_true = tf.cast(tf.argmax(y_true, axis=-1), tf.int64)
+    y_pred = tf.cast(tf.argmax(y_pred, axis=-1), tf.int64)
     true_negatives = tf.reduce_sum(tf.round(tf.clip_by_value((1 - y_true) * (1 - y_pred), 0, 1)))
     possible_negatives = tf.reduce_sum(tf.round(tf.clip_by_value(1 - y_true, 0, 1)))
     
@@ -41,8 +41,8 @@ def sensitivity(y_true, y_pred):
     # n = tf.keras.metrics.TruePositives(threshold)
     # n.update_state(y_true,y_pred)
     # tp = m.result().numpy()
-    y_true = tf.argmax(y_true, axis=-1)
-    y_pred = tf.argmax(y_pred, axis=-1)
+    y_true = tf.cast(tf.argmax(y_true, axis=-1), tf.int64)
+    y_pred = tf.cast(tf.argmax(y_pred, axis=-1), tf.int64)
     true_positives = tf.reduce_sum(tf.round(tf.clip_by_value(y_true * y_pred, 0, 1)))
     possible_positives = tf.reduce_sum(tf.round(tf.clip_by_value(y_true, 0, 1)))
